# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ProfilFinder.at API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  isAdmin: boolean;
  status: string;
  createdAt: string;
}

export interface ChangeEmailInput {
  newEmail: string;
  password: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface PublicProfile {
  id: number;
  profileNumber: string;
  /** @nullable */
  publicName?: string | null;
  jobTitle: string;
  /** @nullable */
  profession?: string | null;
  industry: string;
  /** @nullable */
  experience?: string | null;
  /** @nullable */
  education?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  languages?: string | null;
  /** @nullable */
  desiredJob?: string | null;
  federalState: string;
  /** @nullable */
  city?: string | null;
  availability: string;
  /** @nullable */
  employmentType?: string | null;
  driverLicense?: boolean;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  cvText?: string | null;
  /** @nullable */
  profileImage?: string | null;
  isAvailable: boolean;
  isPublic?: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ProfileListResponse {
  profiles: PublicProfile[];
  total: number;
  page: number;
  limit: number;
}

export interface ProfileInput {
  firstName?: string;
  lastName?: string;
  publicName?: string;
  jobTitle: string;
  profession?: string;
  industry: string;
  experience?: string;
  education?: string;
  skills?: string;
  languages?: string;
  desiredJob?: string;
  federalState: string;
  city?: string;
  availability: string;
  employmentType?: string;
  driverLicense?: boolean;
  description: string;
  cvText?: string;
  profileImage?: string;
  isAvailable?: boolean;
  isPublic?: boolean;
  phoneNumber?: string;
  contactEmail?: string;
  address?: string;
}

export interface ProfileUpdate {
  firstName?: string;
  lastName?: string;
  publicName?: string;
  jobTitle?: string;
  profession?: string;
  industry?: string;
  experience?: string;
  education?: string;
  skills?: string;
  languages?: string;
  desiredJob?: string;
  federalState?: string;
  city?: string;
  availability?: string;
  employmentType?: string;
  driverLicense?: boolean;
  description?: string;
  cvText?: string;
  profileImage?: string;
  isAvailable?: boolean;
  isPublic?: boolean;
  phoneNumber?: string;
  contactEmail?: string;
  address?: string;
}

export interface AdminProfile {
  id: number;
  profileNumber: string;
  userId?: number;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  publicName?: string | null;
  jobTitle: string;
  /** @nullable */
  profession?: string | null;
  industry: string;
  /** @nullable */
  experience?: string | null;
  /** @nullable */
  education?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  languages?: string | null;
  /** @nullable */
  desiredJob?: string | null;
  federalState: string;
  /** @nullable */
  city?: string | null;
  availability: string;
  /** @nullable */
  employmentType?: string | null;
  driverLicense?: boolean;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  cvText?: string | null;
  /** @nullable */
  profileImage?: string | null;
  isAvailable: boolean;
  isPublic?: boolean;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface AdminProfileListResponse {
  profiles: AdminProfile[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminProfileUpdate {
  jobTitle?: string;
  profession?: string;
  industry?: string;
  experience?: string;
  education?: string;
  skills?: string;
  languages?: string;
  desiredJob?: string;
  federalState?: string;
  city?: string;
  availability?: string;
  employmentType?: string;
  driverLicense?: boolean;
  description?: string;
  cvText?: string;
  isAvailable?: boolean;
  isPublic?: boolean;
  adminNotes?: string;
}

export interface CompanyRequestInput {
  profileNumber: string;
  companyName: string;
  contactPerson: string;
  email: string;
  phoneNumber?: string;
  message: string;
}

export interface CompanyRequest {
  id: number;
  /** @nullable */
  profileId?: number | null;
  profileNumber: string;
  companyName: string;
  contactPerson: string;
  email: string;
  /** @nullable */
  phoneNumber?: string | null;
  message: string;
  status: string;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface RequestStatusUpdate {
  status: string;
}

export interface AdminStats {
  totalProfiles: number;
  activeProfiles: number;
  totalRequests: number;
  newRequests: number;
  totalUsers: number;
}

export interface AdminUser {
  id: number;
  email: string;
  isAdmin: boolean;
  status: string;
  createdAt: string;
}

export interface UserStatusUpdate {
  status: string;
}

export interface IndustryStat {
  industry: string;
  count: number;
}

export type GetProfilesParams = {
search?: string;
industry?: string;
federalState?: string;
availability?: string;
profession?: string;
page?: number;
limit?: number;
};

export type GetAdminProfilesParams = {
search?: string;
industry?: string;
federalState?: string;
availability?: string;
page?: number;
limit?: number;
};