# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProfilFinder.at API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /user:
    get:
      operationId: getUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /change-email:
    post:
      operationId: changeEmail
      tags: [auth]
      summary: Change email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeEmailInput"
      responses:
        "200":
          description: Email changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/profiles:
    get:
      operationId: getAdminProfiles
      tags: [admin]
      summary: Get all profiles (admin, with private data)
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: industry
          schema:
            type: string
        - in: query
          name: federalState
          schema:
            type: string
        - in: query
          name: availability
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: All profiles with private data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfileListResponse"
  /admin/profiles/{id}:
    get:
      operationId: getAdminProfile
      tags: [admin]
      summary: Get a single profile with full data (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full profile with private data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
    patch:
      operationId: updateAdminProfile
      tags: [admin]
      summary: Update profile (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
    delete:
      operationId: deleteAdminProfile
      tags: [admin]
      summary: Delete profile (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/company-requests/{id}/status:
    patch:
      operationId: updateCompanyRequestStatus
      tags: [admin]
      summary: Update company request status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyRequest"
...
  /admin/users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [admin]
      summary: Update user status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, email, isAdmin, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        isAdmin:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    ChangeEmailInput:
      type: object
      required: [newEmail, password]
      properties:
        newEmail:
          type: string
          format: email
        password:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
...
    AdminProfile:
      type: object
      required: [id, profileNumber, jobTitle, industry, federalState, availability, isAvailable, createdAt]
      properties:
        id:
          type: integer
        profileNumber:
          type: string
        userId:
          type: integer
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        publicName:
          type: ["string", "null"]
        jobTitle:
          type: string
        profession:
          type: ["string", "null"]
        industry:
          type: string
        experience:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        languages:
          type: ["string", "null"]
        desiredJob:
          type: ["string", "null"]
        federalState:
          type: string
        city:
          type: ["string", "null"]
        availability:
          type: string
        employmentType:
          type: ["string", "null"]
        driverLicense:
          type: boolean
        description:
          type: ["string", "null"]
        cvText:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isPublic:
          type: boolean
        phoneNumber:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    AdminProfileListResponse:
      type: object
      required: [profiles, total, page, limit]
      properties:
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/AdminProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminProfileUpdate:
      type: object
      properties:
        jobTitle:
          type: string
        profession:
          type: string
        industry:
          type: string
        experience:
          type: string
        education:
          type: string
        skills:
          type: string
        languages:
          type: string
        desiredJob:
          type: string
        federalState:
          type: string
        city:
          type: string
        availability:
          type: string
        employmentType:
          type: string
        driverLicense:
          type: boolean
        description:
          type: string
        cvText:
          type: string
        isAvailable:
          type: boolean
        isPublic:
          type: boolean
        adminNotes:
          type: string
...
    CompanyRequest:
      type: object
      required: [id, profileNumber, companyName, contactPerson, email, message, status, createdAt]
      properties:
        id:
          type: integer
        profileId:
          type: ["integer", "null"]
        profileNumber:
          type: string
        companyName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phoneNumber:
          type: ["string", "null"]
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    RequestStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
...
    AdminUser:
      type: object
      required: [id, email, isAdmin, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        isAdmin:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
...