# lib/api-zod/src/generated/types/adminProfileListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ProfilFinder.at API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminProfile } from './adminProfile';

export interface AdminProfileListResponse {
  profiles: AdminProfile[];
  total: number;
  page: number;
  limit: number;
}