# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ProfilFinder.at API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  email: string;
  isAdmin: boolean;
  status: string;
  createdAt: Date;
}