# lib/api-zod/src/generated/types/changeEmailInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ProfilFinder.at API
 * OpenAPI spec version: 0.1.0
 */

export interface ChangeEmailInput {
  newEmail: string;
  password: string;
}