# lib/api-zod/src/generated/types/companyRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ProfilFinder.at API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyRequest {
  id: number;
  /** @nullable */
  profileId?: number | null;
  profileNumber: string;
  companyName: string;
  contactPerson: string;
  email: string;
  /** @nullable */
  phoneNumber?: string | null;
  message: string;
  status: string;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
}